support highlight text
